# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateWorkerRequest:
    properties:
      base_url:
        example: http://localhost:8080
        type: string
      concurrency:
        example: 10
        minimum: 1
        type: integer
      default_delay:
        example: 0
        type: integer
      default_retry_count:
        example: 3
        type: integer
      default_timeout:
        example: 30
        type: integer
      queues:
        additionalProperties:
          type: integer
        type: object
      redis_addr:
        example: redis://localhost:6379/0
        type: string
      worker_name:
        example: my-worker
        type: string
    required:
    - concurrency
    - queues
    - worker_name
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: 错误信息
        type: string
    type: object
  dto.HeartbeatResponse:
    properties:
      heartbeat_at:
        type: string
      status:
        example: ok
        type: string
      worker_name:
        example: my-worker
        type: string
    type: object
...
  dto.RegisterWorkerRequest:
    properties:
      base_url:
        example: http://localhost:8080
        type: string
      concurrency:
        example: 10
        minimum: 1
        type: integer
      default_delay:
        example: 0
        type: integer
      default_retry_count:
        example: 3
        type: integer
      default_timeout:
        example: 30
        type: integer
      queues:
        additionalProperties:
          type: integer
        type: object
      redis_addr:
        example: redis://localhost:6379/0
        type: string
      worker_name:
        example: my-worker
        type: string
    required:
    - concurrency
    - queues
    - worker_name
    type: object
...
  dto.WorkerListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/workers.Config'
        type: array
    type: object
  dto.WorkerResponse:
    properties:
      worker:
        $ref: '#/definitions/workers.Config'
    type: object
  dto.WorkerStatsResponse:
    properties:
      stats: {}
    type: object
  dto.WorkerTimeSeriesResponse:
    properties:
      timeseries: {}
    type: object
...
  workers.Config:
    properties:
      base_url:
        type: string
      concurrency:
        type: integer
      default_delay:
        description: seconds
        type: integer
      default_retry_count:
        type: integer
      default_timeout:
        description: seconds
        type: integer
      is_enabled:
        type: boolean
      last_heartbeat_at:
        type: string
      queues:
        additionalProperties:
          type: integer
        description: queue_name -> weight
        type: object
      redis_addr:
        type: string
      worker_name:
        type: string
    type: object
host: localhost:28080
info:
  contact:
    name: TaskPM Support
  description: 分布式任务管理系统 - 基于 Asynq 和 PostgreSQL 的任务调度平台
  license:
    name: MIT
  title: TaskPM API
  version: 1.0.0
paths:
  /api/v1/workers:
    get:
      description: 获取所有已注册的 Worker 配置列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.WorkerListResponse'
      summary: 获取 Worker 列表
      tags:
      - Workers
    post:
      consumes:
      - application/json
      description: 创建新的 Worker 或更新已存在的 Worker 配置
      parameters:
      - description: Worker 配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateWorkerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.WorkerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 创建或更新 Worker
      tags:
      - Workers
  /api/v1/workers/{worker_name}:
    get:
      description: 根据 worker_name 获取 Worker 配置详情
      parameters:
      - description: Worker 名称
        in: path
        name: worker_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.WorkerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取 Worker 详情
      tags:
      - Workers
  /api/v1/workers/{worker_name}/heartbeat:
    post:
      description: 更新指定 Worker 的心跳时间
      parameters:
      - description: Worker 名称
        in: path
        name: worker_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HeartbeatResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 更新 Worker 心跳
      tags:
      - Workers
  /api/v1/workers/{worker_name}/stats:
    get:
      description: 获取指定 Worker 的任务统计信息
      parameters:
      - description: Worker 名称
        in: path
        name: worker_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.WorkerStatsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取 Worker 统计信息
      tags:
      - Workers
  /api/v1/workers/{worker_name}/timeseries:
    get:
      description: 获取指定 Worker 的时间序列统计数据
      parameters:
      - description: Worker 名称
        in: path
        name: worker_name
        required: true
        type: string
      - default: 24
        description: 统计小时数
        in: query
        name: hours
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.WorkerTimeSeriesResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取 Worker 时间序列统计
      tags:
      - Workers
  /api/v1/workers/register:
    post:
      consumes:
      - application/json
      description: Worker SDK 自动注册接口
      parameters:
      - description: Worker 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterWorkerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.WorkerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 注册 Worker
      tags:
      - Workers
...
schemes:
- http
- https
swagger: "2.0"